requirements.txt
matplotlib==3.10.6
numpy==2.3.2
pandas==2.3.2
peft==0.17.1
Pillow==11.3.0
scikit_learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentence_transformers==5.1.0
torch==2.3.0+cu121
tqdm==4.67.1
transformers==4.56.0.dev0
